controlnotas/requirements.txt
Flask
dash
pandas
plotly
mysql-connector-python
gunicorn